warn for variable length arrays not provably <= 256 bytes
This is a friendlier alternative to introducing -Wvla and attempting to
completely purge variable-length arrays. It warns on any VLA where the
compiler cannot prove that the maximum size isn't larger than 256 bytes.

256 bytes is used rather than the frame size limit warning, since this
doesn't take into account other local variables including multiple VLAs.

The warning is only active for -O2 and above since it depends on
optimization for analysis. For example, this won't warn:

	void f(u8 *a);

	static size_t calculate_size() {
		return AES_MAX_KEY_SIZE;
	}

	void foo() {
		u8 a[calculate_size()];
		f(a);
	}

The warning can be silenced by adding an explicit size check preventing
an overly large VLA from being allocated. Using BUG_ON(size < limit) is
an option to communicate that it can never happen to the compiler, but
that will still warn for !CONFIG_BUG where it's a no-op. For example:

	void f(u8 *a);

	static size_t check_limit(size_t n) {
		BUG_ON(n > AES_MAX_KEY_SIZE);
		return n;
	}

	void foo(size_t n) {
		u8 key[check_limit(n)];
		f(key);
	}

The VLA can be moved to a block or an inner function to implement a
check able to report an error or fall back to dynamic allocation. It's a
bit more painful than it would be with C99 variable declaration.

Signed-off-by: Daniel Micay <danielmicay@gmail.com>